fix(runner): own workspace lifecycle
