+++ packages/desktop-shell/src/desktop-runner.mts
@@ -1,12 +1,10 @@
-import fs from "node:fs";
-import os from "node:os";
-import path from "node:path";
 import { spawn, type ChildProcess } from "node:child_process";
+import fs from "node:fs";
+import path from "node:path";
 import {
   attachProcessStderr,
   reserveLocalPort,
   resolveBunBinary,
-  runCommand,
   stopChildProcess,
   waitForPort,
 } from "./node-utils.mjs";
@@ -28,20 +26,13 @@ type PromptRunnerTaskArgs = {
   sessionId: string;
 };
 
+type DeleteRunnerWorkspaceArgs = {
+  workspaceDirectory: string;
+};
+
 type RunnerProcess = {
   baseUrl: string;
   child: ChildProcess;
-};
-
-type PreparedWorktree = {
-  branchName: string;
-  defaultDirectory: string;
-  directory: string;
-};
-
-type RepoWorkspacePaths = {
-  defaultDirectory: string;
-  repoRoot: string;
 };
 
 type AppRunnerController = {
@@ -50,15 +41,21 @@ type AppRunnerController = {
     sessionId: string;
     workspaceDirectory: string;
   }>;
+  deleteRunnerWorkspace: (args: DeleteRunnerWorkspaceArgs) => Promise<void>;
   promptRunnerTask: (args: PromptRunnerTaskArgs) => Promise<void>;
   stop: () => Promise<void>;
 };
 
-type EnsureAssistantSessionResponse = {
+type CreateAssistantSessionResponse = {
   sessionId: string;
+  workspaceDirectory: string;
 };
 
 type PromptTaskAssistantSessionResponse = {
+  ok: boolean;
+};
+
+type DeleteWorkspaceResponse = {
   ok: boolean;
 };
 
@@ -80,28 +77,21 @@ export function createDesktopRunnerController({
     }
 
     const runner = await ensureRunner();
-    const worktree = prepareSessionWorktree(repoUrl, trimmedTitle);
+    const payload = await postRunnerJson<CreateAssistantSessionResponse>(
+      `${runner.baseUrl}/assistant/session/create`,
+      {
+        model: DEFAULT_OPENCODE_MODEL,
+        provider: DEFAULT_OPENCODE_PROVIDER,
+        repoUrl,
+        taskTitle: trimmedTitle,
+      },
+    );
 
-    try {
-      const payload = await postRunnerJson<EnsureAssistantSessionResponse>(
-        `${runner.baseUrl}/assistant/session/ensure`,
-        {
-          directory: worktree.directory,
-          model: DEFAULT_OPENCODE_MODEL,
-          provider: DEFAULT_OPENCODE_PROVIDER,
-          taskTitle: trimmedTitle,
-        },
-      );
-
-      return {
-        runnerType: "local-worktree",
-        sessionId: payload.sessionId,
-        workspaceDirectory: worktree.directory,
-      };
-    } catch (error) {
-      cleanupFailedWorktree(worktree);
-      throw error;
-    }
+    return {
+      runnerType: "local-worktree",
+      sessionId: payload.sessionId,
+      workspaceDirectory: payload.workspaceDirectory,
+    };
   }
 
   async function promptRunnerTask(args: PromptRunnerTaskArgs): Promise<void> {
@@ -123,6 +113,16 @@ export function createDesktopRunnerController({
     if (!payload.ok) {
       throw new Error("Local runner task prompt did not complete successfully");
     }
+  }
+
+  async function deleteRunnerWorkspace({
+    workspaceDirectory,
+  }: DeleteRunnerWorkspaceArgs): Promise<void> {
+    const runner = await ensureRunner();
+
+    await postRunnerJson<DeleteWorkspaceResponse>(`${runner.baseUrl}/workspace/delete`, {
+      workspaceDirectory,
+    });
   }
 
   async function stop(): Promise<void> {
@@ -190,200 +190,10 @@ export function createDesktopRunnerController({
 
   return {
     createRunnerSession,
+    deleteRunnerWorkspace,
     promptRunnerTask,
     stop,
   };
-}
-
-function prepareSessionWorktree(repoUrl: string, title: string): PreparedWorktree {
-  const workspace = resolveRepoWorkspacePaths(repoUrl);
-  const defaultBranch = ensureDefaultCheckout(repoUrl, workspace);
-  const identifier = nextWorktreeIdentifier(workspace.repoRoot, title);
-  const directory = path.join(workspace.repoRoot, identifier);
-  const branchName = `runner/${identifier}`;
-
-  runCommand(
-    "git",
-    [
-      "-C",
-      workspace.defaultDirectory,
-      "worktree",
-      "add",
-      "-b",
-      branchName,
-      directory,
-      defaultBranch,
-    ],
-    `Failed to create a worktree at ${directory}`,
-  );
-
-  return {
-    branchName,
-    defaultDirectory: workspace.defaultDirectory,
-    directory,
-  };
-}
-
-function cleanupFailedWorktree(worktree: PreparedWorktree): void {
-  try {
-    runCommand(
-      "git",
-      ["-C", worktree.defaultDirectory, "worktree", "remove", "--force", worktree.directory],
-      undefined,
-    );
-  } catch {}
-
-  try {
-    runCommand("git", ["-C", worktree.defaultDirectory, "branch", "-D", worktree.branchName]);
-  } catch {}
-}
-
-function ensureDefaultCheckout(repoUrl: string, workspace: RepoWorkspacePaths): string {
-  fs.mkdirSync(workspace.repoRoot, { recursive: true });
-
-  if (!fs.existsSync(workspace.defaultDirectory)) {
-    cloneDefaultCheckout(repoUrl, workspace.defaultDirectory);
-  } else if (!fs.existsSync(path.join(workspace.defaultDirectory, ".git"))) {
-    throw new Error(`Managed checkout exists without git metadata: ${workspace.defaultDirectory}`);
-  }
-
-  runCommand(
-    "git",
-    ["-C", workspace.defaultDirectory, "fetch", "origin", "--prune"],
-    `Failed to fetch the default checkout in ${workspace.defaultDirectory}`,
-  );
-
-  const defaultBranch = resolveDefaultBranch(workspace.defaultDirectory);
-
-  runCommand(
-    "git",
-    ["-C", workspace.defaultDirectory, "checkout", defaultBranch],
-    `Failed to checkout ${defaultBranch} in ${workspace.defaultDirectory}`,
-  );
-
-  runCommand(
-    "git",
-    ["-C", workspace.defaultDirectory, "pull", "--ff-only", "origin", defaultBranch],
-    `Failed to fast-forward ${defaultBranch} in ${workspace.defaultDirectory}`,
-  );
-
-  return defaultBranch;
-}
-
-function cloneDefaultCheckout(repoUrl: string, defaultDirectory: string): void {
-  fs.mkdirSync(path.dirname(defaultDirectory), { recursive: true });
-
-  const repoSlug = parseRepoSlug(repoUrl);
-  runCommand(
-    "gh",
-    ["repo", "clone", repoSlug, defaultDirectory],
-    `Failed to clone ${repoSlug} into ${defaultDirectory}`,
-  );
-}
-
-function resolveDefaultBranch(defaultDirectory: string): string {
-  const reference = runCommand(
-    "git",
-    ["-C", defaultDirectory, "symbolic-ref", "--short", "refs/remotes/origin/HEAD"],
-    `Failed to resolve the default branch for ${defaultDirectory}`,
-  )
-    .trim()
-    .replace(/^origin\//u, "");
-
-  if (!reference) {
-    throw new Error(`Failed to resolve the default branch for ${defaultDirectory}`);
-  }
-
-  return reference;
-}
-
-function nextWorktreeIdentifier(repoRoot: string, title: string): string {
-  const titleSlug = slugifyIdentifier(title);
-  const timestamp = Math.floor(Date.now() / 1_000);
-
-  for (let attempt = 0; attempt < 100; attempt += 1) {
-    const identifier =
-      attempt === 0 ? `${timestamp}-${titleSlug}` : `${timestamp}-${titleSlug}-${attempt}`;
-
-    if (!fs.existsSync(path.join(repoRoot, identifier))) {
-      return identifier;
-    }
-  }
-
-  throw new Error(`Failed to allocate a unique worktree directory in ${repoRoot}`);
-}
-
-function slugifyIdentifier(value: string): string {
-  let slug = "";
-  let lastWasSeparator = false;
-
-  for (const character of value) {
-    if (/[a-z0-9]/iu.test(character)) {
-      slug += character.toLowerCase();
-      lastWasSeparator = false;
-      continue;
-    }
-
-    if (!lastWasSeparator) {
-      slug += "-";
-      lastWasSeparator = true;
-    }
-  }
-
-  const trimmed = slug.replace(/^-+|-+$/gu, "");
-  return trimmed || "session";
-}
-
-function resolveRepoWorkspacePaths(repoUrl: string): RepoWorkspacePaths {
-  const repoName = parseRepoName(repoUrl);
-  const repoRoot = path.join(resolveRunnerRoot(), repoName);
-
-  return {
-    defaultDirectory: path.join(repoRoot, "default"),
-    repoRoot,
-  };
-}
-
-function resolveRunnerRoot(): string {
-  return path.join(os.homedir(), "clanki");
-}
-
-function parseRepoSlug(repoUrl: string): string {
-  const normalized = normalizeRepoReference(repoUrl);
-  let repoPath = normalized;
-
-  if (repoPath.startsWith("https://github.com/")) {
-    repoPath = repoPath.slice("https://github.com/".length);
-  } else if (repoPath.startsWith("git@github.com:")) {
-    repoPath = repoPath.slice("git@github.com:".length);
-  } else if (repoPath.startsWith("ssh://git@github.com/")) {
-    repoPath = repoPath.slice("ssh://git@github.com/".length);
-  }
-
-  const segments = repoPath.split("/").filter(Boolean);
-  if (segments.length !== 2) {
-    throw new Error(`Unsupported GitHub repository URL: ${repoUrl}`);
-  }
-
-  return `${segments[0]}/${segments[1]}`;
-}
-
-function parseRepoName(repoUrl: string): string {
-  const repoSlug = parseRepoSlug(repoUrl);
-  const repoName = repoSlug.split("/")[1];
-
-  if (!repoName) {
-    throw new Error(`Unsupported GitHub repository URL: ${repoUrl}`);
-  }
-
-  return repoName;
-}
-
-function normalizeRepoReference(repoUrl: string): string {
-  return repoUrl
-    .trim()
-    .replace(/\/+$/u, "")
-    .replace(/\.git$/u, "");
 }
 
 async function isRunnerHealthy(baseUrl: string): Promise<boolean> {
