+++ packages/desktop-shell/src/preload.mts
@@ -4,6 +4,9 @@ contextBridge.exposeInMainWorld("clankiDesktop", {
   createRunnerSession(title: string, repoUrl: string) {
     return ipcRenderer.invoke("desktop-runner:create-session", { repoUrl, title });
   },
+  deleteRunnerWorkspace(workspaceDirectory: string) {
+    return ipcRenderer.invoke("desktop-runner:delete-workspace", { workspaceDirectory });
+  },
   promptRunnerTask(args: {
     backendBaseUrl: string;
     callbackToken: string;
